refactor : 책 카운트 쿼리 분리
