+++ core/src/main/java/com/dnd/sbooky/core/book/MemberBookRepositoryCustom.java
@@ -1,17 +1,18 @@
 package com.dnd.sbooky.core.book;
 
+import com.dnd.sbooky.core.book.dto.FindBookDTO;
 import com.dnd.sbooky.core.book.dto.FindBookDetailsDTO;
-import com.dnd.sbooky.core.book.dto.FindBooksDTO;
+import java.util.List;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
 public interface MemberBookRepositoryCustom {
 
-    FindBooksDTO findMemberBookByMemberIdAndReadStatus(Long memberId, ReadStatus readStatus);
+    List<FindBookDTO> findMemberBookByMemberIdAndReadStatus(Long memberId, ReadStatus readStatus);
 
     FindBookDetailsDTO findBookDetails(Long memberBookId);
 
     boolean checkBookExist(Long memberId, String title, String author);
 
-    long findCompletedBooks(Long memberId);
+    long countMemberBooks(Long memberId, ReadStatus readStatus);
 }
