+++ core/src/main/java/com/dnd/sbooky/core/book/MemberBookRepositoryImpl.java
@@ -1,8 +1,7 @@
 package com.dnd.sbooky.core.book;
 
+import com.dnd.sbooky.core.book.dto.FindBookDTO;
 import com.dnd.sbooky.core.book.dto.FindBookDetailsDTO;
-import com.dnd.sbooky.core.book.dto.FindBooksDTO;
-import com.dnd.sbooky.core.book.dto.FindBooksDTO.FindBookDTO;
 import com.dnd.sbooky.core.member.QMemberEntity;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
@@ -19,27 +18,23 @@ public class MemberBookRepositoryImpl implements MemberBookRepositoryCustom {
     private final QMemberEntity member = QMemberEntity.memberEntity;
 
     @Override
-    public FindBooksDTO findMemberBookByMemberIdAndReadStatus(Long memberId, ReadStatus readStatus) {
+    public List<FindBookDTO> findMemberBookByMemberIdAndReadStatus(
+            Long memberId, ReadStatus readStatus) {
 
-        List<FindBookDTO> findBookDTOList =
-                queryFactory
-                        .select(
-                                Projections.constructor(
-                                        FindBookDTO.class,
-                                        memberBook.id,
-                                        book.title,
-                                        book.author,
-                                        book.thumbnailUrl,
-                                        memberBook.readStatus))
-                        .from(memberBook)
-                        .join(memberBook.bookEntity, book)
-                        .where(memberBook.memberEntity.id.eq(memberId), readStatusEqual(readStatus))
-                        .orderBy(memberBook.id.desc())
-                        .fetch();
-
-        long totalBookCount = countMemberBooks(memberId, null);
-
-        return new FindBooksDTO(totalBookCount, findBookDTOList);
+        return queryFactory
+                .select(
+                        Projections.constructor(
+                                FindBookDTO.class,
+                                memberBook.id,
+                                book.title,
+                                book.author,
+                                book.thumbnailUrl,
+                                memberBook.readStatus))
+                .from(memberBook)
+                .join(memberBook.bookEntity, book)
+                .where(memberBook.memberEntity.id.eq(memberId), readStatusEqual(readStatus))
+                .orderBy(memberBook.id.desc())
+                .fetch();
     }
 
     @Override
@@ -77,11 +72,7 @@ public class MemberBookRepositoryImpl implements MemberBookRepositoryCustom {
     }
 
     @Override
-    public long findCompletedBooks(Long memberId) {
-        return countMemberBooks(memberId, ReadStatus.COMPLETED);
-    }
-
-    private long countMemberBooks(Long memberId, ReadStatus readStatus) {
+    public long countMemberBooks(Long memberId, ReadStatus readStatus) {
         return queryFactory
                 .select(memberBook.count())
                 .from(memberBook)
