+++ core/src/main/java/com/dnd/sbooky/core/book/dto/FindBookDTO.java
@@ -1,0 +1,6 @@
+package com.dnd.sbooky.core.book.dto;
+
+import com.dnd.sbooky.core.book.ReadStatus;
+
+public record FindBookDTO(
+        Long id, String title, String author, String thumbnailUrl, ReadStatus readStatus) {}
